refactor: Move Dialog component to dialog.tsx
